Trade Analysis Capstone Project 1/spark_airflow/requirements.txt
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
py4j==0.10.9.5
pyarrow==9.0.0
pyspark==3.3.0
requests==2.28.1

